requirements.txt
torch
torchvision
pandas
openpyxl
numpy
Pillow
scikit-learn
tqdm
requests
matplotlib
seaborn
umap-learn
